# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: WTS – 모의투자 Web Trading System API
  version: 0.0.1-SNAPSHOT
  description: '모의투자 Web Trading System API.


    클라이언트는 각 서비스를 직접 호출하지 않고 Gateway(http://localhost:8080)로만 접근한다.

    토큰은 `POST /api/users/mock-login` 으로 발급받는다.


    주문 API는 `Idempotency-Key` 헤더가 필수다. 같은 키로 같은 내용을

    재요청하면 주문이 새로 생기지 않고 기존 결과가 돌아온다.


    모든 응답에 `X-Trace-Id` 헤더가 붙는다. 에러 본문의 `traceId`와 같은 값이다.

    '
servers:
- url: http://localhost:8080
  description: Gateway
security:
- bearerAuth: []
...
paths:
...
  /api/trading/orders:
    get:
      tags:
      - 주문
      summary: 주문 목록
      description: 최근 주문부터 돌려준다. status=ACCEPTED로 미체결 주문만 볼 수 있다.
      operationId: list
      parameters:
      - name: status
        in: query
        description: 주문 상태로 거르기. ACCEPTED면 미체결만. 생략하면 전체
        required: false
        schema:
          type: string
          enum:
          - RECEIVED
          - VALIDATED
          - ACCEPTED
          - FILLED
          - CANCELLED
          - REJECTED
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
        '400':
          description: 'Bad Request


            - `VALIDATION_FAILED` — 요청 값이 올바르지 않습니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 'Unauthorized


            - `UNAUTHORIZED` — 인증이 필요합니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-error-codes:
      - UNAUTHORIZED
      - VALIDATION_FAILED
    post:
      tags:
      - 주문
      summary: 주문 접수
      description: 시장가(MARKET)는 최신 현재가로 즉시 전량 체결되어 FILLED로 돌아온다. 지정가(LIMIT)는 BUY면 현재가 ≤ 지정가, SELL이면 현재가
        ≥ 지정가일 때 즉시 체결되고, 아니면 ACCEPTED로 남아 시세가 조건에 도달하면 자동 체결된다. 체결가는 지정가가 아니라 체결 시점 현재가다. 거절되면 에러 응답이
        오고, 거절된 주문도 REJECTED로 주문 내역에 남는다. 같은 Idempotency-Key로 같은 내용을 다시 보내면 주문을 새로 만들지 않고 기존 주문을 그대로 201로
        돌려준다.
      operationId: place
      parameters:
      - name: Idempotency-Key
        in: header
        description: 주문마다 새로 만든 고유 키 (UUID 권장, 1~64자). 같은 키 재요청은 기존 주문을 돌려준다
        required: true
        schema:
          type: string
        example: 550e8400-e29b-41d4-a716-446655440000
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceOrderRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: 'Bad Request


            - `VALIDATION_FAILED` — 요청 값이 올바르지 않습니다.

            - `INSUFFICIENT_BALANCE` — 주문 가능한 예수금이 부족합니다.

            - `INSUFFICIENT_POSITION` — 매도 가능한 수량이 부족합니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 'Unauthorized


            - `UNAUTHORIZED` — 인증이 필요합니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 'Conflict


            - `DUPLICATE_ORDER_REQUEST` — 같은 Idempotency-Key로 다른 주문을 요청했습니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 'Service Unavailable


            - `MARKET_PRICE_UNAVAILABLE` — 현재가를 조회할 수 없습니다.

            - `MARKET_PRICE_STALE` — 현재가가 오래되어 주문을 처리할 수 없습니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-error-codes:
      - UNAUTHORIZED
      - VALIDATION_FAILED
      - MARKET_PRICE_UNAVAILABLE
      - MARKET_PRICE_STALE
      - INSUFFICIENT_BALANCE
      - INSUFFICIENT_POSITION
      - DUPLICATE_ORDER_REQUEST
  /api/trading/orders/{orderId}:
    get:
      tags:
      - 주문
      summary: 주문 상세
      description: 주문 하나를 조회한다. 내 주문이 아니면 404다.
      operationId: one
      parameters:
      - name: orderId
        in: path
        description: 주문 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 2
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: 'Bad Request


            - `VALIDATION_FAILED` — 요청 값이 올바르지 않습니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 'Unauthorized


            - `UNAUTHORIZED` — 인증이 필요합니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 'Not Found


            - `ORDER_NOT_FOUND` — 주문을 찾을 수 없습니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-error-codes:
      - UNAUTHORIZED
      - VALIDATION_FAILED
      - ORDER_NOT_FOUND
    delete:
      tags:
      - 주문
      summary: 주문 취소
      description: 미체결(ACCEPTED) 주문을 취소하고 묶인 예수금·수량을 푼다. 취소된 주문(status=CANCELLED)을 돌려준다. 이미 FILLED / CANCELLED
        / REJECTED인 주문은 409다.
      operationId: cancel
      parameters:
      - name: orderId
        in: path
        description: 주문 ID
        required: true
        schema:
          type: integer
          format: int64
        example: 2
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: 'Bad Request


            - `VALIDATION_FAILED` — 요청 값이 올바르지 않습니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 'Unauthorized


            - `UNAUTHORIZED` — 인증이 필요합니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 'Not Found


            - `ORDER_NOT_FOUND` — 주문을 찾을 수 없습니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 'Conflict


            - `INVALID_ORDER_STATE` — 현재 주문 상태에서는 처리할 수 없습니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-error-codes:
      - UNAUTHORIZED
      - VALIDATION_FAILED
      - INVALID_ORDER_STATE
      - ORDER_NOT_FOUND
...
  /api/users/me/watchlist:
    get:
      tags:
      - 관심종목
      summary: 관심종목 목록
      description: 담은 순서대로 돌려준다.
      operationId: list
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WatchlistItemResponse'
        '401':
          description: 'Unauthorized


            - `UNAUTHORIZED` — 인증이 필요합니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-error-codes:
      - UNAUTHORIZED
  /api/users/me/watchlist/{symbol}:
    post:
      tags:
      - 관심종목
      summary: 관심종목 추가
      description: 이미 담긴 종목이어도 성공이다(같은 요청을 여러 번 보내도 결과가 같다). 종목이 실제로 존재하는지는 확인하지 않는다. 최대 개수(기본 50개)를 넘으면
        거절한다.
      operationId: add
      parameters:
      - name: symbol
        in: path
        description: 종목코드 (6자리)
        required: true
        schema:
          type: string
        example: 005930
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistItemResponse'
        '400':
          description: 'Bad Request


            - `VALIDATION_FAILED` — 요청 값이 올바르지 않습니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 'Unauthorized


            - `UNAUTHORIZED` — 인증이 필요합니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-error-codes:
      - UNAUTHORIZED
      - VALIDATION_FAILED
    delete:
      tags:
      - 관심종목
      summary: 관심종목 삭제
      description: 담겨 있지 않은 종목을 빼도 성공이다. 본문 없이 204를 돌려준다.
      operationId: remove
      parameters:
      - name: symbol
        in: path
        description: 종목코드 (6자리)
        required: true
        schema:
          type: string
        example: 005930
      responses:
        '204':
          description: No Content
        '401':
          description: 'Unauthorized


            - `UNAUTHORIZED` — 인증이 필요합니다.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-error-codes:
      - UNAUTHORIZED
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        code:
          type: string
          description: 에러 코드
          example: INSUFFICIENT_BALANCE
        message:
          type: string
          description: 사람이 읽을 수 있는 메시지
          example: 주문 가능한 예수금이 부족합니다.
        traceId:
          type: string
          description: 추적 ID. 응답 헤더 X-Trace-Id와 같다
          example: effc7d36-ebd7-46ae-9ad8-9ea215d500b5
        timestamp:
          type: string
          format: date-time
          description: 발생 시각 (UTC)
          example: '2026-09-22T08:12:44.101Z'
...
    OrderResponse:
      type: object
      properties:
        orderId:
          type: integer
          format: int64
          description: 주문 ID
          example: 2
        symbol:
          type: string
          description: 종목코드
          example: 005930
        side:
          type: string
          description: BUY 매수 / SELL 매도
          enum:
          - BUY
          - SELL
          example: BUY
        orderType:
          type: string
          description: MARKET 시장가 / LIMIT 지정가
          enum:
          - MARKET
          - LIMIT
          example: LIMIT
        quantity:
          type: integer
          format: int64
          description: 주문 수량 (주)
          example: 10
        limitPrice:
          type: number
          description: 지정가 (원). 시장가 주문이면 필드가 없다
          example: 80000
        filledQuantity:
          type: integer
          format: int64
          description: 체결된 수량 (주)
          example: 0
        status:
          type: string
          description: 주문 상태. FILLED 전량 체결 / ACCEPTED 미체결(예수금·수량 묶임) / CANCELLED 취소 / REJECTED 거절
          enum:
          - RECEIVED
          - VALIDATED
          - ACCEPTED
          - FILLED
          - CANCELLED
          - REJECTED
          example: ACCEPTED
        rejectReason:
          type: string
          description: '거절 사유 에러 코드 (예: INSUFFICIENT_BALANCE). REJECTED 주문에만 있다'
        createdAt:
          type: string
          format: date-time
          description: 주문 시각 (UTC)
          example: '2026-09-22T08:12:44.101Z'
        updatedAt:
          type: string
          format: date-time
          description: 마지막 상태 변경 시각 (UTC)
          example: '2026-09-22T08:12:44.101Z'
    PlaceOrderRequest:
      type: object
      properties:
        symbol:
          type: string
          description: 종목코드. 6자리 숫자
          example: 005930
          pattern: \d{6}
        side:
          type: string
          description: BUY 매수 / SELL 매도
          enum:
          - BUY
          - SELL
          example: BUY
        orderType:
          type: string
          description: MARKET 시장가 / LIMIT 지정가
          enum:
          - MARKET
          - LIMIT
          example: LIMIT
        quantity:
          type: integer
          format: int64
          description: 주문 수량 (주). 1 ~ 1,000,000
          example: 10
          maximum: 1000000
          minimum: 1
        limitPrice:
          type: number
          description: 지정가 (원). LIMIT 주문에 필수, 0보다 커야 한다. MARKET 주문이면 무시한다
          example: 80000
          minimum: 0
      required:
      - orderType
      - quantity
      - side
      - symbol
...
    WatchlistItemResponse:
      type: object
      properties:
        symbol:
          type: string
          description: 종목코드
          example: 005930
        createdAt:
          type: string
          format: date-time
          description: 담은 시각 (UTC)
          example: '2026-09-22T08:12:44.101Z'
  securitySchemes:
    bearerAuth:
      type: http
      description: '`POST /api/users/mock-login` 이 발급한 accessToken'
      scheme: bearer
...